[Fix] #26 로그인 연동완료 +  수강 강의 조회 기능 추가

+++ LoveTriangle-LMS/src/main/java/com/module1/crud/course/model/service/CourseService.java
@@ -30,7 +30,7 @@ public class CourseService {
         }
     }
 
-    public List<CourseDTO> findMyCourses(Long userId) throws SQLException {
+    public List<CourseDTO> findMyCourses(int userId) throws SQLException {
         Connection con = JDBCTemplate.getConnection();
 
         List<CourseDTO> myCourses = courseDAO.findMyCourses(userId);
